Initial check-in of Proxy Tester v0.8
If you would like to see the original repository where I started
development on this project then please see the project on
sourceforge.  It will be kept around for archival purposes.

http://sourceforge.net/projects/proxytest/

The original documentation wiki will be kept around for archival
purposes.

http://sourceforge.net/apps/mediawiki/proxytest/

This project is the brain child of Sam Gleske (me).  I thought it
would be really cool to bounce around different proxies like they
do in the movies when connecting to a website.  So the tested proxies
and generated WPAD file do just that :).